file name changed, lombok added, api route chaged
