+++ backend/src/main/java/com/laligaStats/searchPlayers/controller/PlayerController.java
@@ -1,7 +1,7 @@
-package com.laligaStats.soccerproject.controller;
+package com.laligaStats.searchPlayers.controller;
 
-import com.laligaStats.soccerproject.entity.PlayerEntity;
-import com.laligaStats.soccerproject.service.PlayerService;
+import com.laligaStats.searchPlayers.entity.PlayerEntity;
+import com.laligaStats.searchPlayers.service.PlayerService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -12,7 +12,7 @@ import java.util.List;
 
 @Slf4j
 @RestController
-@RequestMapping("/api")
+@RequestMapping("/api/players")
 public class PlayerController {
 
     private final PlayerService playerService;
