+++ backend/src/main/java/com/laligaStats/searchPlayers/entity/PlayerEntity.java
@@ -1,0 +1,27 @@
+package com.laligaStats.searchPlayers.entity;
+
+import jakarta.persistence.*;
+import lombok.Data;
+
+@Entity
+@Table(name = "player_data")
+@Data
+public class PlayerEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    Long id;
+
+    String player_name;
+    String nation;
+    String positions;
+    Integer age;
+    Integer matches_played;
+    Double goals;
+    Double assists;
+    Double penalty_goals;
+    Double yellow_card;
+    Double red_card;
+    Double expected_goals;
+    Double expected_assists;
+    String team_name;
+}
