+++ backend/src/main/java/com/laligaStats/searchPlayers/repository/PlayerRepository.java
@@ -1,6 +1,6 @@
-package com.laligaStats.soccerproject.repository;
+package com.laligaStats.searchPlayers.repository;
 
-import com.laligaStats.soccerproject.entity.PlayerEntity;
+import com.laligaStats.searchPlayers.entity.PlayerEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PlayerRepository extends JpaRepository<PlayerEntity, Long> {
