+++ backend/src/test/java/com/laligaStats/searchPlayers/LaLigaStatsApplicationTests.java
@@ -1,4 +1,4 @@
-package com.laligaStats.soccerproject;
+package com.laligaStats.searchPlayers;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
